feat: 통합 검색 폼을 UnifiedSearch 컴포넌트로 분리
- DashboardScreen 내 인라인 아이콘(SearchIcon/MicIcon/SparkleIcon) 및 form 블록 추출

- app/components/UnifiedSearch.tsx 신규 생성, 단일 책임 원칙 적용

- DashboardScreen에서 ~70줄 제거 후 UnifiedSearch 로 교체
